+++ cli/src/pcluster/config/cluster_config.py
@@ -2566,14 +2566,14 @@ class _CommonQueue(BaseQueue):
             )
 
 
-class AllocationStrategy(Enum):
-    """Define supported allocation strategies."""
-
-    LOWEST_PRICE = "lowest-price"
-    CAPACITY_OPTIMIZED = "capacity-optimized"
-    PRICE_CAPACITY_OPTIMIZED = "price-capacity-optimized"
-    PRIORITIZED = "prioritized"
-    CAPACITY_OPTIMIZED_PRIORITIZED = "capacity-optimized-prioritized"
+# class AllocationStrategy(Enum):
+#     """Define supported allocation strategies."""
+#
+#     LOWEST_PRICE = "lowest-price"
+#     CAPACITY_OPTIMIZED = "capacity-optimized"
+#     PRICE_CAPACITY_OPTIMIZED = "price-capacity-optimized"
+#     PRIORITIZED = "prioritized"
+#     CAPACITY_OPTIMIZED_PRIORITIZED = "capacity-optimized-prioritized"
 
 
 class SlurmQueue(_CommonQueue):
