+++ cli/src/pcluster/config/common.py
@@ -27,6 +27,14 @@ from pcluster.validators.s3_validators import UrlValidator
 
 LOGGER = logging.getLogger(__name__)
 
+class AllocationStrategy(Enum):
+    """Define supported allocation strategies."""
+
+    LOWEST_PRICE = "lowest-price"
+    CAPACITY_OPTIMIZED = "capacity-optimized"
+    PRICE_CAPACITY_OPTIMIZED = "price-capacity-optimized"
+    PRIORITIZED = "prioritized"
+    CAPACITY_OPTIMIZED_PRIORITIZED = "capacity-optimized-prioritized"
 
 class CapacityType(Enum):
     """Enum to identify the type compute supported by the queues."""
